refactor: split nplikes
