+++ src/awkward/_connect/rdataframe/from_rdataframe.py
@@ -71,7 +71,7 @@ def from_rdataframe(data_frame, columns):
     def empty_buffers(cpp_buffers_self, names_nbytes):
         buffers = {}
         for item in names_nbytes:
-            buffers[item.first] = ak._nplikes.numpy.empty(item.second)
+            buffers[item.first] = numpy.empty(item.second)
             cpp_buffers_self.append(
                 item.first,
                 buffers[item.first].ctypes.data_as(ctypes.POINTER(ctypes.c_ubyte)),
