+++ src/awkward/_nplikes/__init__.py
@@ -1,0 +1,6 @@
+from awkward._nplikes.array_module import ArrayModuleNumpyLike  # noqa: F401
+from awkward._nplikes.cupy import Cupy  # noqa: F401
+from awkward._nplikes.dispatch import nplike_of, to_nplike  # noqa: F401
+from awkward._nplikes.jax import Jax  # noqa: F401
+from awkward._nplikes.numpy import Numpy  # noqa: F401
+from awkward._nplikes.numpylike import ArrayLike, NumpyLike, NumpyMetadata  # noqa: F401
